Create README.md
It is a Web based application that is functional for college students in a accessing the library facility online.
A user is prompted to sign up and log in with their admission number. The user will get to choose the type of book to acquire, the department the student comes from and the block the student is going to sit in the library together with the seat.
With this application the user will be able to know which seats are empty and how many are free without even going to the library.
The user will also search and know which books are still in the shelves and if the book the user is trying to have has been booked, taken or not taken.
The student will also be able to apply for a library membership card in the app.